refactoring: login validation configs.
