+++ src/modules/auth/ui/containers/LogInFormContainer.tsx
@@ -1,10 +1,10 @@
 import { FormProvider } from 'react-hook-form';
 import { LogInForm } from '@/modules/forms';
-import useLogInForm from '@/modules/auth/hooks/ui/useLogInForm';
 import Stack from '@mui/material/Stack';
 import SectionWrapper from '@/shared/ui/SectionWrapper';
 import LogInHeader from '@/modules/auth/ui/components/LogInHeader';
 import GuestCredentials from '@/modules/auth/ui/components/GuestCredentials';
+import useLogInForm from '@/features/login/model/useLogInForm';
 
 const LogInFormContainer = () => {
   const {
